requirements.txt
selenium>=4.0.0
webdriver-manager
playwright
pytest
pyyaml
